chore: remove manuscript sync TS runtime import and refresh predev manifest (pb: pb-f5d)
- Added a shared plain-JS manuscript parser helper so the sync script and TS wrapper use the same parsing rules
- Reworked scripts/sync-manuscripts.js to read proofs directly and avoid importing lib/manuscript.ts at runtime
- Extended predev to run sync:manuscripts alongside critique asset sync
- Reverified manuscript sync, typecheck, and production build after the cleanup
